requirements.txt
absl-py==0.7.1
appdirs==1.4.4
asgiref==3.3.4
astor==0.8.0
astroid==2.4.2
astunparse==1.6.3
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.9.3
bleach==3.1.4
bs4==0.0.1
cachetools==4.1.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.3
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
djangorestframework==3.9.2
entrypoints==0.3
fake-useragent==0.1.11
future==0.18.2
gast==0.2.2
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.30.0
gunicorn==20.1.0
h5py==2.9.0
idna==2.8
importlib-metadata==1.7.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
jedi==0.14.0
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.0
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.6.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.17.0
oauthlib==3.1.0
opencv-python==4.3.0.36
opt-einsum==3.2.1
pandas==0.24.2
pandocfilters==1.4.2
parse==1.19.0
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.2.0
progressbar==2.5
progressbar2==3.46.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
psycopg2==2.9.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.2.2
Pygments==2.4.2
pylint==2.5.3
pyparsing==2.4.0
pyppeteer==0.2.6
pyquery==1.4.3
pyrsistent==0.15.3
pytesseract==0.3.4
python-dateutil==2.8.0
python-utils==2.4.0
pytz==2019.1
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.0.2
q==2.6
qtconsole==4.5.1
requests==2.22.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.21.2
scipy==1.5.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.1
sympy==1.4
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.1
torch==1.5.1
torch-utils==0.1.2
torchvision==0.6.1
tornado==6.0.3
tqdm==4.47.0
traitlets==4.3.2
typed-ast==1.4.1
typing-extensions==3.10.0.0
urllib3==1.25.11
utils==1.0.1
w3lib==1.22.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==9.1
Werkzeug==0.15.4
whitenoise==5.3.0
widgetsnbextension==3.5.0
wrapt==1.11.2
zipp==3.1.0
